Extract the logic in hook
